Move cv2_utils to a separate module
